+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Added Generic Profile implementation
 * Added @phpstan-require-implements constraints
 
 ## v0.4.3 (2024-08-21)